🔥 : 기존 trading table 삭제하고 폴더 정리
shadcn/ui 사용한 파일 삭제 (리액트 테이블로 변경함)
