+++ src/app/(main)/trading/page.tsx
@@ -2,10 +2,8 @@ import type { MiniCardPT } from "@/common/components/card"
 import { TitleCard, CardComponent, MiniCard } from "@/common/components/card"
 import { LineComponent } from "@/features/realtime/components/chart"
 import { VolumeChart } from "@/features/realtime/components/chart/trading-volume-chart"
-import { TradingDataTable2sss } from "@/features/realtime/components/realtime-trading-table"
-import { TradingTable } from "@/features/realtime/components/realtime-trading-table/data-table"
-import { TradingVolumeCards } from "@/features/realtime/components/slide-card"
 import { TradingDataTable } from "@/features/realtime/components/trading-table"
+import { TradingVolumeCards } from "@/features/realtime/components/slide-card"
 
 export default function Trading() {
 	const miniCardDatas = [
@@ -63,9 +61,7 @@ export default function Trading() {
 				</div>
 			</div>
 			<div className="flex w-full justify-center">
-				<CardComponent>
-					<TradingDataTable2sss />
-				</CardComponent>
+				<TradingDataTable />
 			</div>
 		</div>
 	)
