+++ src/features/realtime/components/trading-table/Tcolumn.tsx
