--- src/features/realtime/components/trading-table/data.tsx
@@ -1,126 +1,0 @@
-import type { DataTablePT } from "./Tcolumns"
-export default function getData(): DataTablePT[] {
-	return [
-		{
-			id: "trade001",
-			tradeTime: "2023-12-04 12:00",
-			plantName: "서울 태양광 발전소",
-			volume: "500",
-			bidPrice: "70",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade002",
-			tradeTime: "2023-12-04 13:00",
-			plantName: "부산 풍력 발전소",
-			volume: "300",
-			bidPrice: "65",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade003",
-			tradeTime: "2023-12-04 14:30",
-			plantName: "인천 태양광 발전소",
-			volume: "200",
-			bidPrice: "72",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade004",
-			tradeTime: "2023-12-04 15:00",
-			plantName: "대구 지열 발전소",
-			volume: "150",
-			bidPrice: "68",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade005",
-			tradeTime: "2023-12-04 16:00",
-			plantName: "광주 태양광 발전소",
-			volume: "600",
-			bidPrice: "75",
-			matchingStatus: "matched",
-		},
-
-		{
-			id: "trade006",
-			tradeTime: "2023-12-04 17:00",
-			plantName: "대전 태양광 발전소",
-			volume: "450",
-			bidPrice: "69",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade007",
-			tradeTime: "2023-12-04 18:00",
-			plantName: "울산 풍력 발전소",
-			volume: "350",
-			bidPrice: "71",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade008",
-			tradeTime: "2023-12-04 19:00",
-			plantName: "세종 태양광 발전소",
-			volume: "400",
-			bidPrice: "67",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade009",
-			tradeTime: "2023-12-04 20:00",
-			plantName: "제주 풍력 발전소",
-			volume: "300",
-			bidPrice: "64",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade010",
-			tradeTime: "2023-12-04 21:00",
-			plantName: "강릉 태양광 발전소",
-			volume: "500",
-			bidPrice: "72",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade011",
-			tradeTime: "2023-12-04 22:00",
-			plantName: "포항 지열 발전소",
-			volume: "250",
-			bidPrice: "66",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade012",
-			tradeTime: "2023-12-04 23:00",
-			plantName: "평택 태양광 발전소",
-			volume: "550",
-			bidPrice: "74",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade013",
-			tradeTime: "2023-12-05 00:00",
-			plantName: "수원 풍력 발전소",
-			volume: "600",
-			bidPrice: "73",
-			matchingStatus: "unmatched",
-		},
-		{
-			id: "trade014",
-			tradeTime: "2023-12-05 01:00",
-			plantName: "창원 태양광 발전소",
-			volume: "400",
-			bidPrice: "68",
-			matchingStatus: "matched",
-		},
-		{
-			id: "trade015",
-			tradeTime: "2023-12-05 02:00",
-			plantName: "춘천 태양광 발전소",
-			volume: "350",
-			bidPrice: "70",
-			matchingStatus: "unmatched",
-		},
-	]
-}
